apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
redis==5.0.8
google-generativeai==0.8.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
httpx==0.27.0
pgvector==0.3.2
itsdangerous==2.2.0
